backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
hdrezka>=4.0.0
python-dotenv>=1.0.0
